[build] Add missing deps, fix build race
I uncovered this issue while playing with GN internals
and messing up the timing of actions.
I think previously the source_set was "losing" the race
to the actions that were generating the code and because
of that this was never a problem. But since my change
made actions a lot slower, clang started winning the
race and would then complain that it couldn't find the
generated files.

Change-Id: I46352d3267a4849b27c252cc4a7b15841f75e413
Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/mesa/+/437475
Reviewed-by: Petr Hosek <phosek@google.com>