[gn] add mesa_python_stdout_to_file_action template
This makes the call-sites much more readable and compact.

Also replace the python GN wrapper script with a bash version.
The old python script was using the no-longer supported python2.7,
and python is really over-kill for what it was doing.

Also use the in-tree python interpreter rather than
relying on whatever `python` happens to be in the path.
This makes the build more hermetic.

Bug: 62771
Bug: 60924
Change-Id: Icc7bd10243b15671971ab690853b951ad774cc2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/mesa/+/445078
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: John Rosasco <rosasco@google.com>